--- doc_generator/generate_dummy.py
@@ -1,5 +1,0 @@
-# import models.cadquery.rack_leg
-
-def generate_docs(config, config_hash):
-    print("I am modifying config files and running ExSource-Tools...")
-    print("I will put my finished artifacts in a directory marked with the config hash: " + str(config_hash))
