+++ doc_generator/mechanical/components/cadquery/rack_leg.py
